requirements.txt
openai
Pillow
moviepy
